revamp PacketRegistry (#3179)
* revamp PacketRegistry

* remove BIDIRECTIONAL

* Move packets to PacketVanilla

* Use ? extends T

* feat: Update to 26.2

Also simplify using Map instead of ClassValue, and don't use packets method anymore.

---------
